split into mobile and desktop view (#40)
